requirements.txt
Django~=2.0.6
django-cors-headers==2.4.0
djangorestframework==3.8.2
mysqlclient==1.3.13
requests==2.20.1